bug(PayInAdvance) - Don't loose precision when converting timestamp (#3274)
## Description

This small changes ensure that we find subscriptions that start on the
same *second* as the event timestamp. previously the conversion to an
integer would drop the milliseconds from the float, resulting in a
timestamp that starts that the start of the second.

Due to this we could no longer find subscriptions starting in the same
second. (as the event now suddenly falls before the start date of the
subscription).

This change fixes that and converts the timestamp to a float.