feat(data): Add graphql endpoint for revenue streams plans (#3278)
**Context**

A lot of users ain’t using our analytics feature because it’s not
detailed enough. Thus, we are revamping the Analytics dashboard.

**Description**

This pull request introduces a new feature to query revenue streams
plans of an organization using GraphQL.
It requests the `lago-data` api service and returns revenue streams per
plans for an organization.