feat(data): Add mrrs plans resolver (#3293)
**Context**

A lot of users ain’t using our analytics feature because it’s not
detailed enough. Thus, we are revamping the Analytics dashboard.

**Description**

This pull request introduces a new feature to query mrrs plans of an
organization using GraphQL.
It requests the `lago-data` api service and returns mrrs per plan for an
organization.