feat(data): Add usages invoiced resolver (#3306)
**Context**

A lot of users ain’t using our analytics feature because it’s not
detailed enough. Thus, we are revamping the Analytics dashboard.

**Description**

This pull request introduces a new feature to query usages invoiced of
an organization using GraphQL.
It requests the `lago-data` api service and returns usages invoiced for
an organization.