fix(dunning): adapt refund logic with dunning (#3308)
## Context

Refund logic does not work with dunning campaign

## Description

Currently credit note refund logic works only if payment intent is
created with the initial invoice and if it is successful.

However, with dunning campaign, initial payment can be failed, but there
can be separate payment request which covers invoice.

This case was not supported and with refund we were linking invalid
failed payment.

This PR fixes described case.