feat(kafka): Add in advance charge consumer, to replace sidekiq hack (#3291)
## Context

This PR is related to https://github.com/getlago/lago/pull/474

## Description

The main goal is to add a Kafka consumer responsible for handling events
charged in advance. It replace the the "hacky" approach used in the
current post-processor implementation that is pushing Sidekiq jobs from
the go service.

The consumer relies on the Karafka gem, making it easy to add new
consumer in the future.

For now the `EventsChargedInAdvanceConsumer` will only enqueue a
`Events::PayInAdvanceJob` for every message received in the topic,
making it easy to keep the simple setup without kafka/clickhouse