fix(preauth): Remove unnecessary amount_capturable check (#3319)
## Context

This is a beta feature only enabled for few customer

## Description

When a payment pre-authorization is performed, it's possible to get a
successful response from Stripe but that for some reason the
`amount_capturable` isn't the expected amount. I originally added this
extra check because I think it's not what we expect but in the end, it's
probably best to let the rest of the code run and try the actually
payment.

Some payment will work anyway, some will probably fail but it's better
than cancelling the subscription creation while the payment could have
gone through.