feat(regroup_fee): Use current month boundaries for AdvanceCharges invoices (#3327)
## Context

`advance_charges` invoices will regroup all **non invoiceable fees paid
within the last period**.

Invoice billing periods are normally related to the fee periods but in
this case, you can have a fee created in June paid only in November so
we want to recompute the periods.

The second issue is that the `DateService` is designed to work active or
"recently terminated" subscription. In this case we try to compute dates
for a fees long after the subscription is terminated which results in
wrong dates (`to` date being before `from` date for instance). This bug
is NOT fixed. Even if it was, the Date service would still compute date
related to the subscription when the fee is created which is **not what
we want in this specific case**.

## Description

This fixed is inspired by this first attempt by @vincent-pochet :
https://github.com/getlago/lago-api/pull/3285

Note that this invoice only has charge fees by design, so subscription
fee.

For `advance_charges` and only for these invoices, I created a new
service to create the `InvoiceSubscription` model. We grab the most
recent subscription to compute correct boundaries and apply them to all
InvoiceSubscriptions.


This invoice is special! See other corner case recently fixed:
https://github.com/getlago/lago-api/pull/3153