misc(retry-jobs): Adjust jitter to increase spread between repeating jobs (#3340)
## Context

We had a batch of 600 failed jobs trying to update 3 wallets which lead
to jobs failing with `ActiveRecord::StaleObjectError: Attempted to
update a stale object: Wallet.` exception.
We're trying to increase `jitter` for job retries to order to increase
spread between jobs retries and see if that it's enough to mitigate the
issue.

This value should increase spread by 10 times on 3rd attempt.

## Description

Increase job jitter