chore(eu-vat): Update eu_vat_rates.json (#3352)
This file was copy-pasted from the original repo:
https://github.com/ibericode/vat-rates

~This might be communicated to users and clients (via changelog).~

This file will create the taxes in DB when it's "EU Tax Management"
integration is installed but the rate are not updated automatically.

> [!IMPORTANT]  
> Notice that compared to the [original
file](https://github.com/ibericode/vat-rates/blob/master/vat-rates.json),
we manually remove `GB`. I'm not sure why but it was done last time:
https://github.com/getlago/lago-api/pull/1570