feat(dunning): Add BCC emails addresses (#3354)
## Description

When creating a dunning campaign, you can now add a list of email
addresses to pass in the `bcc` field. This can help accounting team
keeping track of payment requests.

In GraphQL, the field is an array of string, each string is a valid
email address. It's NOT a single string containing comma-separated email
addresses (like we might do in other places).

```rb
["email@test.com", "another@test.com"] 
🆚 
"email@test.com,another@test.com"
```

The dunning campaign is attached to the payment request so I only had to
check it in the Mailer. The rest is basically model + gql boilerplate.