feat(payments): Save payment method on successful payment (stripe only) (#3345)
## Context

This PR reintroduce the work reverted in #3282 

I originally wanted to freeze the stripe version before re-introducing
this feature but:
* I want to upgrade Stripe before we freeze the version
* you can't [update webhook
endpoint](https://docs.stripe.com/webhooks/versioning), you need to
handle both at some point, so let's handle both now.

## Description

We want to save some payment method details for a successful payment.
This info will be reused in the Payment Receipt feature.

It's done with the webhook only because it needs to be done with the
webhook for all payment happening outside Lago (like invoice payment url
or "customer balance"). Instead of doing it multple times, we always
rely on webhooks.

The card details are not part of all webhooks depending on the version
so we call the Stripe API to get the details of the payment method.

We call the service synchronously because this will allow to dispatch
jobs that rely on this info (like payment receipt 😬)
We could easily make this async if needed.

### Caching ?

We could introduce some cache because the payment method ID is unique
but I'm not sure how often we'd hit the cache. I think a lot of
customers have one or 2 invoice a month. Only few have a lot payments
per months. We could be writing a lot of stuff to redis and by the time
we need it (next month) it was removed.

```ruby
Rails.cache.fetch("pm-data-#{provider_payment_method_id}") do
  retrieve_stripe_payment_method_data
end
```