misc(GraphQL): add `has_*` attributes to billable_metric object as replacement of `*_count` attributes (#3379)
These methods will replace `*_count` methods as this ones are DB
expensive and we don't need them anymore. It is enough to know if
billable metric has related objects nor the count of them, and this is
more performant.

`*_count` attributes are deprecated and will be deleted after frontend
does not require them anymore.