feat(stripe): Update payment method in a job (#3381)
## Description

I was hesitant to make it async but it's better for retries.

Note that `update_payment_status` might raise a not found error if the
webhook arrived before the `payment._provider_payment_id` is updated in
the DB, which will retry the HandleEventJob.

So when `UpdatePaymentMethodDataJob` is dispatched, we're guaranteed
that payment is updated... until the implementation of
`update_payment_status` changes 😄