misc(GraphQL): add `has_*` attributes to plan object as replacement of `*_count` attributes
These methods will replace `*_count` methods as this ones are DB
expensive and we don't need them anymore. It is enough to know if
plan or its children plans has related objects nor the count of them,
also, this is performant.

`*_count` attributes are deprecated and will be deleted after frontend
does not require them anymore.