fix(payment): Dont update payment method if payment not found (#3389)
## Description

See https://github.com/getlago/lago-api/pull/3381

If the initial service updating the payment method (which might also
create payment if not found doesn't return a payment), ignore the next
job.

What I got wrong initially is that I assumed if no payment was found in
the end, the service was raising an error. It doesn't.


https://github.com/getlago/lago-api/blob/953b92a96de74ab613479dd4db1d5e85977d9d47/app/services/invoices/payments/stripe_service.rb#L16-L28


https://github.com/getlago/lago-api/blob/953b92a96de74ab613479dd4db1d5e85977d9d47/app/services/payment_requests/payments/stripe_service.rb#L32-L42