fix(payments): Fix payment of a one time payment request (#3441)
## Context

There was a bug when a payment for payment request was initiated via
clicking a payment link in an email.
In that case we were creating a new payment instead of finding an
existing one by payment provider id.
That new payment was not created due to a unique index on
`provider_payment_id` and `payment_provider_id`.

## Description

This PR fixes the described issue by fixing
`Invoices::Payments::StripeService#update_payment_status` method
implementation.