fix(fee): Add repeatable read isolation for in advance fees creation (#3444)
## Context

An issue was reported by a customer, related to pay in advance fees
related to a percentage charge model with a per transaction max amount.

The in advance charge always performs two queries for the aggregation,
one including the billed event and a second one excluding it. The final
amount is then the difference of these two amounts. On top of that the
per transaction max/min amount is also querying the DB to fetch the
discreet values of each event in order to apply the min/max logic on
each of them.

If the billing of an event is started and a new event is ingested in the
mean time, it could lead to race condition were the new event is taken
into account in some queries and not in other.

## Description

This PR enforce the transaction isolation to `repeatable_read` during
the aggregation to make sure that the new events cannot interfere with
the result of the aggregation.

To avoid write in the DB and reducing the risk of failure with the
isolated transaction, the service is also refactored to first initialize
the fees in isolation mode and then persist them in a standard
transaction.