fix(invoices): Ensure no organization are in a middle state (#3440)
Add a migration to queue a job.

This job will ensure no organization are in a middle state upgrading
from `per_customer` to `per_organization`, if this is the case, assign
the non self billed with generated number invoices count to the last
invoice. This will ensure next invoice will get the right
organization_sequential_id.

using a job instead of running the data migration within the actual
migration has two benefits;
1. does not block the release with a long running migration.
2. no need to redefine all models scopes and relationships.

with one drawback;
We can not ensure the migration was done on time... but for this case it
should be fine-ish