Change billing_entity_sequential_id assignation strategy
Now we only populate the billing_entity_sequential_id when
document_nubmer is per_billing_enitty. With this change we will be
missing some information but we do this to reduce as much as
possible locking the tables when running billing. The lock will be only
needed when the document numbering strategy is per billing entity. If it
is per_customer we will reduce the locking impact as it wont be applied
at billing_entity level...

fix test rebase