feat(fee): reset timestamps when updating fee.payement_status (#3471)
## Context

Until a Fee is attached to an invoice, its payment_status can be updated
via the API

## Description

If a fee goes from pending to failed, we'll set the `failed_at`. Then,
when going from failed to succeeded, we'll set the `succeeded_at`. If
the fee goes back to failed or pending, the `succeeded_at` field is
still set.

This PR result all 3 timestamps (`failed_at`, `succeeded_at` and
`refunded_at`) so only the timestamp associated with the new
payment_status is set.