chore(spec): Send `customer.vies_check` webhook if error (#3472)
## Description

Adds scenario tests about customer taxes and EU tax management. Ensure a
webhook is sent in case of success or failure.

This ensure the webhook is sent in case of error

When using Valvat, the gem will perform format check via Regex before
calling the API. If the format is invalid, the API isn't called.

In the webhook pay load the `customer.vies_check` hash has multple form.

**Notice that valvat is now autloaded with Bundler. I don't know why it
was done manually (since the gem was added).**

### When valid

The response from the VIES API is returned, like before

### When format is invalid

```json
{
  "valid": false,
  "valid_format": false,
}
```

### When format is valid but number doesn't exist

```json
{
  "valid": false,
  "valid_format": true
}
```


### When the API returned an error (timeout, rate limit and so on)

```json
{
  "valid": false,
  "valid_format": true,
  "error": "The  web service returned the error: <ERROR FROM VIES>"
}
```

The error message prefix is from valvat.