misc(filter): Remove OpenStruct for Query::Filters definition (#3499)
## Context

This PR is part of the global initiative to remove `OpenStruct` from the
codebase

## Description

This PR replaces the definition of `BaseQuery::Filters` defined as an
`OpenStruct` with a new `BaseFilter` class.
In a similar way of the `BaseResult` this new class enforces the
whitelisting of accepted filters, by defining the full list of accepted
arguments.

NOTE: A future pull request will rework this to integrate with
`Queries::*QueryFiltersContract` definitions