Fix(stripe-webhooks): do not process payment on stripe webhook that belongs to another organization (#3511)
## Context

In case when multiple lago organizations are registered within the same
stripe account, we'll have a webhook registered per each organization on
stripe.
But if an event happened on stripe, it will send this event to each
webhook.
Right now we'll find a payment that is sent in the event and will
process it with the new data.
So if the event was sent on 3 webhooks, we'll try to process this event
three times.
Instead, since in the webhook we have the organization_id, it should be
processed only when the related registered in lago payment belongs to
the organization, sent in the webhook

(on the image below you can see logs filtered by having a specific
payment_intent in the payload, sent to different webhooks)
<img width="1123" alt="image"
src="https://github.com/user-attachments/assets/0161d6da-c05f-4e33-8960-3f132cbf3cd8"
/>


## Description

- When processing the payment, if payment is found, so we don't have to
create it, do not process it if it belongs to another organization
- also removed raise on conflicted
`Payments::SetPaymentMethodAndCreateReceiptJob`
- added more default info in the payment metadata when sent to Stripe