feat: Add create_org helper to Rails console (#3521)
## Context

This process usually involves calling the Organizations::CreateService
and then manually creating an admin invite.
To streamline this, we’re adding a new helper method to
Rails::ConsoleMethods.

## Description

This PR adds a create_organization(org_name:, email:) method to the
Rails::ConsoleMethods module. It:
Creates an organization using Organizations::CreateService
Creates an admin invite for the specified email via
Invites::CreateService
Prints the invitation URL in the console

Also returns the organization and invite_url in a hash for further use.