feat(Idempotency): introduce IdempotencyRecord (#3494)
## Description

IdempotencyRecords are a new way to enforce idempotency on the DB level
where you can't have all data in a single table.
(if you can enforce it with a uniqueness constraint, that approach is
better, this should only be used if you need data across tables!)

This PR adds the groundwork for this feature, next step is to use this
for progressive billing.