Feat(multientities): payment receipt pdf use billing_entity_info (#3519)
## Context

when generating payment receipt PDFs we should use the billing_entity,
associated to the payable of this receipt. Payable can be invoice or
payment_request. Invoice belongs to a billing_entity, but
payment_request did not, so we added a relation on payment_request to
the billing_entity through customer

## Description

- updated payment receipt pdfs
- added relation to billing_entity from payment_request through customer