fix(invoice-preview): Handle downgraded subscriptions in invoice previews (#3534)
## Context

Currently we're generating credit note for case when we simulate
subscription downgrade. That was creating issues since for downgrade
case existing subscription always billed for full billing period and
there's nothing to refund.

## Description

Adjust logic in `Invoices::PreviewService` to skip credit note
generation for terminated subscription if it's downgrade case.