misc: Add organization_id field on charges (#3532)
## Context

This PR is part of the epic to add `organization_id` on all tables of
the application

## Description

The current one is adding the field to the `charges` table. For now the
field allows null values. It will be back-filled in a later pull request
and then updated to ensure the presence of a value (not null)