Fix(invoices): Ensure invoices got organization sequential id (#3549)
Queue again the database migration job as it failed in the previous
attempt due to code loading once the job was queued.

This change queues again the job to ensure no organization is missing
the organization sequential id on its last invoice if it changed from
numbering per customer to per organization... very unlikely at this
point.