chore(spec): Ensure all tables have organization_id column (#3565)
## Context

We're adding `organization_id` to all tables.

## Description

In order not to forget to add new `organization_id` to new tables, I'm
adding this spec which queres the db schema.
After all tables are migrated, the expectation will be updated to `.to
be_empty`.

Remove the table name every time you migrate one.