feat(audit_logs): Create Clickhouse::ActivityLog model (#3502)
## Roadmap Task

👉
[https://getlago.canny.io/feature-requests/p/access-audit-logs](https://getlago.canny.io/feature-requests/p/access-audit-logs)

## Context

Our users want to have visibility on what’s happening in Lago.

Examples:
- *A plan and prices have been updated: When? By whom? What?*
- *A subscription has been deleted: When? By whom? What?*

## Description

The goal of this PR is to introduce a new Clickhouse table
`activity_logs` for storing all activity events from our customers.

We decided to put logs on Clickhouse instead of PG.
This is the first type of logs, but we want to add api and security logs
in the future.