feat(audit_logs): Create GraphQL classes for activity log (#3573)
## Roadmap Task

👉
[https://getlago.canny.io/feature-requests/p/access-audit-logs](https://getlago.canny.io/feature-requests/p/access-audit-logs)

## Context
Our users want to have visibility on what’s happening in Lago.

Examples:

A plan and prices have been updated: When? By whom? What?
A subscription has been deleted: When? By whom? What?

## Description

The goal of this PR is to create the a new Graphql classes for
activity_logs, so they can presented on Front.

---------

Co-authored-by: Mario Diniz <mariohenriquediniz@gmail.com>