chore(organization): Add organization_id to customer_metadata table (#3591)
## Context

This PR is part of the epic to add `organization_id` on all tables of
the application

## Description

The current one is adding the field to the `customer_metadata` table.
For now the field allows null values. A job to back-fill it has been
added.
In a later pull request the not null constraint will be added