chore(organization): Add organization_id to billable_metric_filters table (#3593)
This PR is part of the epic to add `organization_id` on all tables of
the application

The current one is adding the field to the `billable_metric_filters`
table.
For now the field allows null values. A job to back-fill it has been
added.
In a later pull request the not null constraint will be added