chore(logs): Log when idempotency record is not unique (#3630)
## Context

IdempotencyRecords were recently added to ensure uniqueness across the
app.

## Description

The goal is to add an entry to the logs so we can start ignoring this
exception when needed. For example, an inbound webhook coming from
payment provider can be ignored if they are processed twice. Yet, I
think we want to add something to the logs.