feat (avalara-integration): Cover void invoice flow (#3638)
## Context

Lago is currently adding integration with tax provider Avalara.

## Description

This PR handles void invoice flow.

When invoice is voided in Lago it needs to be reported in the tax
provider.

The only exception is when Avalara transaction is in the locked state.
In that case, it is not possible to change transaction in any way, so we
need to create negative transaction which will act as the refund.



This one is opened instead of
https://github.com/getlago/lago-api/pull/3604