misc: avoid ActiveRecord::PreparedStatementCacheExpired errors (#3645)
## Context

In some context, when a migration is adding a new column to a table,
Postgres might return a `cached plan must not change result type` in a
transaction because the cached query plan is not more valid. This error
is then raised as a `ActiveRecord::PreparedStatementCacheExpired` by
Rails

See
https://flexport.engineering/avoiding-activerecord-preparedstatementcacheexpired-errors-4499a4f961cf

## Description

This fix is forcing rails to query explicitly every columns in select
instead of using `SELECT *`, so that the cached plan remains valid in
the transaction. The only drawback will be with column removal, but it
should never happen on a used field and would be prevented by
`strong_migrations`