chore(organization): Add organization_id to integration_collection_mappings table (#3673)
## Context

This PR is part of the epic to add `organization_id` on all tables of
the application

## Description

The current one is adding the field to the
`integration_collection_mappings` table.
For now the field allows null values. A job to back-fill it has been
added.
In a later pull request the not null constraint will be added