feat(gql): Add BillableMetricId field on Alert object (#3692)
## Context

In the frontend, we want to filter the list of alert_type and
billable_metric_id to remove those that already exists.

## Description

In GQL, when an object as a realtionship, we typically add this as a
full object. I'm also adding the `billabled_metric_id` on Alert object
which as 2 benefits:

1. we avoid N+1 loading billable metric just to get the ID
1. the response shape is much nicer (no nested hash)


```json
[
  {
    "alertType": "billable_metric_usage_amount",
    "billableMetricId": "b86641de-bcb4-4be7-abf8-57c3f9905213"
  },
  {
    "alertType": "usage_amount",
    "billableMetricId": null
  }
]
```

In order to check if an alert exists, we can loop over the array like
this

```ruby
h.any? { it["alertType"] == "usage_amount" }
```

![CleanShot 2025-05-21 at 12 36
11@2x](https://github.com/user-attachments/assets/2cec5062-b56b-425d-b61f-1eccee60a44a)