feat: Fallback to organization country for EU bank transfers (#3689)
## Context

Stripe only supports a limited list of countries when creating
`eu_bank_transfer` funding instructions. If a customer's country is not
in that list, the request fails with a 400 error.

This feature request aims to allow accepting bank transfers from any
European country by defaulting to the organization if that country is
supported.

## Description

- A fallback mechanism was added when creating Stripe funding
instructions for the type `eu_bank_transfer`.
- If the customer's country is unsupported, the system checks whether
the organization's country is supported.