feat(pricing_group_keys): Add validation logic for all charge models (#3666)
## Context

This PR is part of an epic to add grouped_by feature on all charge
models and not only on the standard one like today.

## Description

This PR is the first step:
- It refactors the validation logic for `Charge` and `ChargeFilter` to
share the logic
- It moves the grouped_by validation logic from the
`Charges::Validators::StandardService` to the
`Charges::Validators::BaseService` so that it can be applied to all
charge models
- It adds the ability to send grouping keys via `pricing_group_keys` or
`grouped_by` to ensure a smooth transition between the two keys.

For now `grouped_by` properties remains whitelisted only on the standard
charge model. This will remain like this until the grouped_by logic is
applied to all charge model