fix(wallet): Optimistic locking for in advance events (#3720)
## Context

This PR is a new attempt to approach the issue of parallel wallet
refresh refresh in the case of in payed in advance events.

## Description

The PR moves fro pessimistic locking (`FOR UPDATE NOWAIT`) to an
optimistic one using the `lock_version` already present on the wallets
table.
An automatic retry logic is added around the wallet decrease logic with
a random wait time between each attempt. It reduce the effect of the
error to the decrease logic only instead of the all event processing
like it is today.
Finally in case of error, the error is re-raised and the entire logic
will be retried by Sidekiq