chore(stripe): Refactor Stripe webhooks (#3718)
## Context

This PR is part of [upgrading
Stripe](https://github.com/getlago/lago-api/pull/3300) (clients + API
version).

## Description

This is extracted from #3300. 

**You cannot update the `api_version` of an existing WebhookEndpoint**.
You must create a new WebhookEndpoint with the new version and delete
the old one once you migrated. This is the best way to not have any
downtime.

https://docs.stripe.com/webhooks/versioning

It means to set the version for all our customers, we'll need to
recreate the endpoint for all organization. It's fine but I'd rather
upgrade to a more recent version before we enforce it in the webhook.

* Replace `PaymentProviders::StripeService.new.refresh_webhook` by
dedicated `RefreshWebhookService`
* Tried to make `PaymentProviders::StripeService.create_or_update` a
little clearer as secrets cannot be updated