feat(stripe): Mark invoice as disputed for PaymentRequest (#3643)
A payment has `payable` which can be an Invoice or a `PaymentRequest`.

When a dispute is lost, Stripe sends a `charge.dispute.closed` webhook
with the payment details included. When attached to an invoice, the
invoice is flagged as disputed but it it was a PaymentRequest, it wasn't
handled.

This PR loops over each invoices of the PaymentRequest and mark them as
disputed.

The Adyen service was also updated and since the test don't mock the
service call, but check the result, they don't need to be updated.


Note that the original `Invoices::LoseDisputeService` is still [used by
GraphQL](https://github.com/getlago/lago-api/blob/b99e8a592fa2d736ce772ff0abff5257fc879543/app/graphql/mutations/invoices/lose_dispute.rb#L19)
when marking an invoice as last.