fix(clickhouse): Add auto-retry logic on SSL reset error (#3733)
## Context

Some time, with Clickhouse queries, a `Errno::ECONNRESET Connection
reset by peer - SSL_connect` is raised.
This happen when the Clickhouse cluster is under eavy load and is scaled
automatically.

## Description

Since these error are intermediary, this PR adds a auto-retry system (up
to 3 retries) to reduce the number of errors in the following processes
(billing, current usage, daily usage...)

NOTE: we should check how to address this at the Clickhouse connector
level, but it is not currently possible to pass an option to enable
`max_retries` on the underlying Net::HTTP client