misc(organization): Exclude remove clickhouse_aggregation column (#3739)
## Description

This PR follows https://github.com/getlago/lago-api/pull/3722.

The `clickhouse_aggregation` was recently removed from the organization
table.
Because the deploy of this change lead to a lot of issues as we are
relying on `ApplicationRecord#ignored_columns` to avoid
`ActiveRecord::PreparedStatementCacheExpired`, this PR mark the removed
column as ignored_columns to reduce the risk of error.

For future columns removal, this approach will be done before pushing
the migration.