ftr(wallet-limitations): add support for limitations in wallet services (#3737)
## Context

Currently wallet credits are applied on all fee types. This feature adds
possibility to limit wallet consumption on specific fee type.

## Description

This PR adds support for new `applies_to` attribute which is the main
object for various limitations. The first limitation will be the one
related to fee type, but in the future `applies_to` will be extended
with more limitations