fix(bug): rescue BaseService::ThrottlingError when calculating Custom… (#3653)
### Problem
`::Invoices::CustomerUsageService` can raise
`BaseService::ThrottlingError` when `apply_taxes` is `true` and
`customer_provider_taxation?` is `true`.

If the controller does not handle this, the user will receive a 500
error.

This PR does not fix the throttling error because its out of our
control, but inform the user to try again later in a nicer way than a
error 500.
#### Sentry error
https://lago.sentry.io/issues/6336411980/?project=6370067

##
### Steps to reproduce
Have an organization with Anrok provider configured
Have a invoice for an Anrok customer

call api GET `customers/<customer_external_id>/current_usage`. 
Apply taxes is `true` by default.
The ThrottlingError may be raised and return a 500 error.
##
### Result
With the rescue in place, the user would get a response like this when
ThrottlingError happens

```
{
"status" => 429,
 "error" => "Too Many Provider Requests",
 "code" => "too_many_provider_requests",
 "error_details" => { 
      "provider_name" => "anrok",
      "message" => "Service anrok is not available. Try again later."
  }
}
 ```