fix(charge): Fix deletion cascade on large plan override (#3765)
## Context

This PR follows https://github.com/getlago/lago-api/pull/3747

Deleting a charge in a plan with a large number of override can cause
Timeout error or lead to a memory error on Redis.

The main reasons are:
- The service will enqueue one job per child charge
- In the service, the soft deletion is done using `discard_all` witch is
not scalable as it is using a DB call per record...


## Description

This PR is refactoring the logic to enqueue only one job responsible of
the soft deletion of all child charges.
The filter deletion is also done by replacing `discard_all` calls with
`update_all(deleted_at:)`

Note: The same logic will have to be applied to both update and creation
of child charges